REFACTOR: move custom Sphinx callbacks to .util submodule
